# swagger.yaml
openapi: 3.0.0
info:
  title: Scientific Publications API
  version: 1.0.0
  description: |
    A comprehensive backend API for analyzing scientific publications with NLP and ML capabilities.
    This system provides endpoints for managing publications, authors, citations, topics, and analytics.
    
    ## Features
    - Publication management with full-text search
    - Author profiles with collaboration networks
    - Citation analysis and relationship mapping
    - Topic modeling and trend analysis
    - Advanced analytics and insights
    - Data import/export capabilities
    - Secure authentication and authorization
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your-jwt-token>`
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Limits vary by endpoint type.
  contact:
    name: API Support
    email: support@scientific-publications.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api
    description: Development server
paths:
...
  /authors/top/{metric}:
    get:
      tags:
        - Authors
      summary: Get top authors by metrics
      description: Retrieves authors ranked by specific metrics
      parameters:
        - name: metric
          in: path
          required: true
          schema:
            type: string
            enum: [h_index, total_citations, total_publications]
          description: Metric to sort by
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Number of results to return
        - name: country
          in: query
          schema:
            type: string
          description: Filter by country
      responses:
        '200':
          description: Top authors retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  authors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Author'
                  metric:
                    type: string
                  total:
                    type: integer
...
  /topics:
    get:
      tags:
        - Topics
      summary: List topics
      description: Retrieves a paginated list of topics with filtering
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: category
          in: query
          schema:
            type: string
            enum: [stem, social_sciences, humanities, arts, interdisciplinary, other]
        - name: field
          in: query
          schema:
            type: string
        - name: trendDirection
          in: query
          schema:
            type: string
            enum: [emerging, growing, stable, declining, resurgent]
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Topics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/Topic'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Topics
      summary: Create new topic
      description: Creates a new topic with hierarchical structure
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicCreate'
      responses:
        '201':
          description: Topic created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
  /topics/{id}:
    get:
      tags:
        - Topics
      summary: Get topic by ID
      description: Retrieves detailed topic information with analytics
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Topic retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  topic:
                    $ref: '#/components/schemas/TopicDetail'
                  analytics:
                    type: object
        '404':
          $ref: '#/components/responses/NotFound'
  /topics/trending/{limit}:
    get:
      tags:
        - Topics
      summary: Get trending topics
      description: Retrieves topics with high growth or popularity
      parameters:
        - name: limit
          in: path
          required: true
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Trending topics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/Topic'
                  total:
                    type: integer
  /topics/search:
    post:
      tags:
        - Topics
      summary: Search topics
      description: Performs full-text search on topics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                filters:
                  type: object
                page:
                  type: integer
                  default: 1
                limit:
                  type: integer
                  default: 20
      responses:
        '200':
          description: Topic search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/Topic'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /topics/top/{metric}:
    get:
      tags:
        - Topics
      summary: Get top topics by metrics
      description: Retrieves topics ranked by specific metrics
      parameters:
        - name: metric
          in: path
          required: true
          schema:
            type: string
            enum: [total_publications, total_citations, popularity_score]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: category
          in: query
          schema:
            type: string
            enum: [stem, social_sciences, humanities, arts, interdisciplinary, other]
      responses:
        '200':
          description: Top topics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/Topic'
                  metric:
                    type: string
                  total:
                    type: integer
  /topics/hierarchy/{rootId}:
    get:
      tags:
        - Topics
      summary: Get topic hierarchy
      description: Retrieves hierarchical structure of topics
      parameters:
        - name: rootId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Topic hierarchy retrieved
          content:
            application/json:
              schema:
                type: object
  /topics/{id}/words:
    get:
      tags:
        - Topics
      summary: Get topic words
      description: Retrieves important words for a topic
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: nWords
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Topic words retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  topic_id:
                    type: string
                  words:
                    type: object
                  count:
                    type: integer
  /topics/{id}/publications:
    get:
      tags:
        - Topics
      summary: Get publications for topic
      description: Retrieves publications associated with a topic
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [publication_year, citations_count, created_at]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Topic publications retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  publications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Publication'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /topics/{id}/related:
    get:
      tags:
        - Topics
      summary: Get related topics
      description: Retrieves topics related to a specific topic
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 20
        - name: relationshipType
          in: query
          schema:
            type: string
            enum: [similar, parent, child, overlapping, competing, complementary]
      responses:
        '200':
          description: Related topics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  topic_id:
                    type: string
                  related_topics:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
  /topics/{id}/validate:
    put:
      tags:
        - Topics
      summary: Update topic validation status
      description: Updates validation status of a topic
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - validation_status
              properties:
                validation_status:
                  type: string
                  enum: [pending, validated, disputed, deprecated]
                validation_notes:
                  type: string
      responses:
        '200':
          description: Validation status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  validation_status:
                    type: string
  /topics/{id}/statistics:
    get:
      tags:
        - Topics
      summary: Get topic statistics
      description: Retrieves detailed statistics for a topic
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Topic statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  topic_id:
                    type: string
                  analytics:
                    type: object

  # Import/Export Routes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of items per page
...
    SortByParam:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [created_at, title, citations_count, publication_year, h_index, total_citations, popularity_score]
      description: Field to sort by
    SortOrderParam:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
...
  schemas:
...
    Institution:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        position:
          type: string
        country:
          type: string
...
    Publication:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        abstract:
          type: string
        publication_year:
          type: integer
        doi:
          type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/AuthorReference'
        journal:
          $ref: '#/components/schemas/Journal'
        citations_count:
          type: integer
        keywords:
          type: array
          items:
            $ref: '#/components/schemas/Keyword'
        created_at:
          type: string
          format: date-time
...
    AuthorReference:
      type: object
      properties:
        author_id:
          type: string
        name:
          type: string
        order:
          type: integer
        affiliation:
          type: string
    Journal:
      type: object
      properties:
        name:
          type: string
        volume:
          type: string
        issue:
          type: string
        pages:
          type: string
        issn:
          type: string
        publisher:
          type: string
    Keyword:
      type: object
      properties:
        term:
          type: string
        weight:
          type: number
        source:
          type: string
...
    Author:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        orcid:
          type: string
        h_index:
          type: integer
        total_citations:
          type: integer
        total_publications:
          type: integer
        affiliations:
          type: array
          items:
            $ref: '#/components/schemas/Institution'
        research_interests:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    Topic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        keywords:
          type: array
          items:
            $ref: '#/components/schemas/Keyword'
        total_publications:
          type: integer
        total_citations:
          type: integer
        popularity_score:
          type: number
        trend_direction:
          type: string
        created_at:
          type: string
          format: date-time
    TopicDetail:
      allOf:
        - $ref: '#/components/schemas/Topic'
        - type: object
          properties:
            parent_topic:
              $ref: '#/components/schemas/Topic'
            children:
              type: array
              items:
                $ref: '#/components/schemas/Topic'
            related_topics:
              type: array
              items:
                $ref: '#/components/schemas/RelatedTopic'
            top_authors:
              type: array
              items:
                $ref: '#/components/schemas/Author'
            top_journals:
              type: array
              items:
                $ref: '#/components/schemas/Journal'
    TopicCreate:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        keywords:
          type: array
          items:
            $ref: '#/components/schemas/Keyword'
        parent_topic_id:
          type: string
    RelatedTopic:
      type: object
      properties:
        topic:
          $ref: '#/components/schemas/Topic'
        relationship_type:
          type: string
        similarity_score:
          type: number
    Pagination:
      type: object
      properties:
        current:
          type: integer
        pages:
          type: integer
        total:
          type: integer
        limit:
          type: integer
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...